refator(structure): refactor log folder structure
