+++ server/api/factory/factory.integration.test.js
@@ -3,7 +3,7 @@ const request = require('supertest');
 const setup = require('../../../tests/setup');
 
 const config = setup.initConfig();
-const logger = require('../../logger/logger')({ config });
+const logger = require('../../log/logger')({ config });
 const middleware = require('../../middleware')({ config, logger });
 const app = require('../../server')({ config, logger, middleware });
 const modelUtil = require('../../../tests/utils/util.model.integration');
