+++ server/middleware/errorHandlingMiddleware.test.js
@@ -1,6 +1,6 @@
 const { initConfig } = require('../../tests/setup');
 
-const logger = require('../logger/logger')({ config: initConfig() });
+const logger = require('../log/logger')({ config: initConfig() });
 const errorHandlingMiddleware = require('./errorHandlingMiddleware')({ logger });
 const response = require('../../tests/utils/response.mock');
